src/main/java/com/daejong/service/impl/CategoryServiceImpl.java
package com.daejong.service.impl;

import com.daejong.dataobject.ProductCategory;
import com.daejong.repository.ProductCategoryRepository;
import com.daejong.service.CategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * 类目
 * Created by Daejong on 2017/9/10.
 */

@Service
public class CategoryServiceImpl implements CategoryService{

    /**
     * 引入DAO
     */
    @Autowired
    private ProductCategoryRepository repository;

    @Override
    public ProductCategory findOne(Integer categoryId) {
        return repository.findOne(categoryId);
    }

    @Override
    public List<ProductCategory> findAll() {
        return repository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryType) {
        return repository.findByCategoryTypeIn(categoryType);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return repository.save(productCategory);
    }
}
